Fix Resolver API use
The "local artifact" is non-null ONLY if result comes from "local"
repository (was once already resolver) and should not be used
as you are not really interested in this.

Just use "artifact" instead, that is non-null even on
first call (when local repo still does not have it and is downloaded
from somewhere).